perf(core): colocate element normalization
Keep createElement and cloneElement normalization together and reuse their argument counts. This reduces core from 4,814 to 4,804 bytes gzip without adding runtime work.
