+++ src/diff/index.js
@@ -691,7 +691,7 @@ function diffElementNodes(
 				// despite the attribute not being present. When the attribute
 				// is missing the progress bar is treated as indeterminate.
 				// To fix that we'll always update it when it is 0 for progress elements
-				(inputValue !== dom[i] || (nodeType === 'progress' && !inputValue))
+				(inputValue !== dom[i] || (nodeType == 'progress' && !inputValue))
 			) {
 				setProperty(dom, i, inputValue, oldProps[i], namespace);
 			}
